fix(vision): preserve SigLIP dtype explicitly

+++ lib/vision/siglip.worker.ts
@@ -47,10 +47,10 @@ async function ensureClassifier() {
   const pipe = (await pipeline(
     "zero-shot-image-classification",
     "Xenova/siglip-base-patch16-224",
-    // Explicit dtype avoids the runtime warning and prevents WebGPU from
-    // defaulting to fp32 for this model. fp16 keeps the local scoring path
-    // lighter on VRAM while preserving the same public scoring contract.
-    { device: "webgpu", dtype: "fp16" } as Parameters<typeof pipeline>[2]
+    // Preserve the previous default precision explicitly so Transformers.js no
+    // longer warns that dtype was unspecified. Do not switch to fp16/q8 without
+    // browser-runtime testing that the model variant exists and scores well.
+    { device: "webgpu", dtype: "fp32" } as Parameters<typeof pipeline>[2]
   )) as unknown as (
     image: Blob,
     labels: string[]
